chore: enable nuget lock files for precise CI caching

+++ EmployeeDirectory.Test/HealthEndpointTests.cs
@@ -1,12 +1,16 @@
+using System;
+using System.IO;
+using System.Net.Http;
+using System.Threading.Tasks;
+using System.Collections.Generic;
 using Microsoft.AspNetCore.Mvc.Testing;
-using Microsoft.VisualStudio.TestPlatform.TestHost;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Configuration.Memory;
 using Xunit;
 
 namespace EmployeeDirectory.Test
 {
-    public class HealthEndpointTests
+    public class HealthEndpointTests : IClassFixture<WebApplicationFactory<Program>>, IDisposable
     {
         private HttpClient _client;
         private string _tempFile;
@@ -19,7 +23,7 @@ namespace EmployeeDirectory.Test
             {
                 builder.ConfigureAppConfiguration((ctx, conf) =>
                 {
-                    var dict = new Dictionary<string, string>
+                    var dict = new Dictionary<string, string?>
                     {
                         ["Storage:Provider"] = "Json",
                         ["Storage:JsonFilePath"] = _tempFile
@@ -41,8 +45,22 @@ namespace EmployeeDirectory.Test
 
         public void Dispose()
         {
-            if (File.Exists(_tempFile))
-                File.Delete(_tempFile);
+            // Dispose the test HTTP client created from the factory
+            _client?.Dispose();
+            _client = null;
+
+            // Remove the temporary storage file if it was created
+            if (!string.IsNullOrEmpty(_tempFile) && File.Exists(_tempFile))
+            {
+                try
+                {
+                    File.Delete(_tempFile);
+                }
+                catch
+                {
+                    // Ignore cleanup failures in the test teardown
+                }
+            }
         }
 
         //[SetUp]
